06-playground-simulation/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6
pyyaml>=6.0
universal-agent-fabric>=0.1.0
openai>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
